+++ lib/razor/command/create_repo.rb
@@ -1,10 +1,45 @@
 # -*- encoding: utf-8 -*-
 class Razor::Command::CreateRepo < Razor::Command
-  authz '%{name}'
+  summary "Create a new repository, from an ISO image or a URL"
+  description <<-EOT
+Create a new repository, which can either contain the content to install a
+node, or simply point to an existing online repository by URL.
+  EOT
 
-  attr  'name',    type: String, required: true
-  attr  'url',     type: URI,    exclude: 'iso-url'
-  attr  'iso-url', type: URI,    exclude: 'url'
+  example <<-EOT
+Create a repository from an ISO image, which will be downloaded and unpacked
+by the razor-server in the background:
+
+    {
+      "name":    "fedora19",
+      "iso-url": "http://example.com/Fedora-19-x86_64-DVD.iso"
+      "task":    "fedora"
+    }
+
+You can also unpack an ISO image from a file *on the server*; this does not
+upload the file from the client:
+    {
+      "name":    "fedora19",
+      "iso-url": "file:///tmp/Fedora-19-x86_64-DVD.iso"
+      "task":    "fedora"
+    }
+
+Finally, you can providing a `url` property when you create the repository;
+this form is merely a pointer to a resource somehwere and nothing will be
+downloaded onto the Razor server:
+
+    {
+      "name": "fedora19",
+      "url":  "http://mirrors.n-ix.net/fedora/linux/releases/19/Fedora/x86_64/os/"
+      "task": "fedora"
+    }
+  EOT
+
+  authz '%{name}'
+  attr  'name',    type: String, required: true, size: 1..250
+  attr  'url',     type: URI,    exclude: 'iso-url', size: 1..1000
+  attr  'iso-url', type: URI,    exclude: 'url', size: 1..1000
+
   object 'task', required: true do
     attr 'name', type: String, required: true
   end
@@ -26,5 +61,11 @@ class Razor::Command::CreateRepo < Razor::Command
     # Finally, return the state (started, not complete) and the URL for the
     # final repo to our poor caller, so they can watch progress happen.
     repo
+  end
+
+  def self.conform!(data)
+    data.tap do |_|
+      data['task'] = { 'name' => data['task'] } if data['task'].is_a?(String)
+    end
   end
 end