+++ lib/razor/command/create_tag.rb
@@ -1,8 +1,24 @@
 # -*- encoding: utf-8 -*-
 
 class Razor::Command::CreateTag < Razor::Command
+  summary "Create a new tag"
+  description <<-EOT
+Create a new tag, and set the rule it will use to match on facts and node
+metadata.
+  EOT
+
+  example <<-EOT
+Create a simple tag:
+
+    {
+      "name": "small",
+      "rule": ["=", ["fact", "processorcount"], "2"]
+    }
+
+  EOT
+
   authz '%{name}'
-  attr  'name', type: String, required: true
+  attr  'name', type: String, required: true, size: 1..Float::INFINITY
   attr  'rule', type: Array
 
   def run(request, data)