+++ lib/razor/command/move_policy.rb
@@ -1,6 +1,22 @@
 # -*- encoding: utf-8 -*-
 
 class Razor::Command::MovePolicy < Razor::Command
+  summary "Change the order that policies are considered when matching against nodes"
+  description <<-EOT
+Policies can be moved before or after specific policies.
+  EOT
+
+  example <<-EOT
+Move a policy before another policy:
+
+    {"name": "policy", "before": "other"}
+
+Move a policy after another policy:
+
+    {"name": "policy", "after": "other"}
+  EOT
+
+
   authz '%{name}'
   attr   'name', type: String, required: true, references: Razor::Data::Policy
 
@@ -24,5 +40,12 @@ class Razor::Command::MovePolicy < Razor::Command
     policy.save
 
     policy
+  end
+
+  def self.conform!(data)
+    data.tap do |_|
+      data['before'] = { 'name' => data['before'] } if data['before'].is_a?(String)
+      data['after'] = { 'name' => data['after'] } if data['after'].is_a?(String)
+    end
   end
 end